# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pet Hotel API
  version: 1.0.0
  description: Backend para reservas de hotel de mascotas (Auth, Users, Clients, Pets, Inventory, Bookings, Payments).
servers:
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BookingItemService:
      type: object
      properties:
        service_id: { type: string, format: uuid }
        qty: { type: integer }
        unit_price: { type: number }
    BookingItem:
      type: object
      properties:
        room_id: { type: string, format: uuid }
        pet_id:  { type: string, format: uuid }
        nightly_rate: { type: number }
        notes: { type: string, nullable: true }
        services:
          type: array
          items: { $ref: '#/components/schemas/BookingItemService' }
    CreateBookingRequest:
      type: object
      required: [customer_user_id, start_date, end_date, items]
      properties:
        customer_user_id: { type: string, format: uuid }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        items:
          type: array
          items: { $ref: '#/components/schemas/BookingItem' }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        customer_user_id: { type: string, format: uuid }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        status: { type: string, enum: [PENDING, CONFIRMED, CANCELLED] }
        subtotal: { type: number }
        taxes: { type: number }
        total: { type: number }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    UpdateBookingStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [PENDING, CONFIRMED, CANCELLED] }
...
paths:
...
  /v1/bookings:
    get:
      tags: [Bookings]
      security: [ { bearerAuth: [] } ]
      summary: Listar reservas
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Booking' }
    post:
      tags: [Bookings]
      security: [ { bearerAuth: [] } ]
      summary: Crear reserva
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookingRequest' }
      responses:
        '201':
          description: Creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
  /v1/bookings/{id}:
    get:
      tags: [Bookings]
      security: [ { bearerAuth: [] } ]
      summary: Detalle reserva con items & services
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Detalle
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
  /v1/bookings/{id}/status:
    patch:
      tags: [Bookings]
      security: [ { bearerAuth: [] } ]
      summary: Cambiar estado reserva
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateBookingStatusRequest' }
      responses:
        '200': { description: OK }
...